Move non-extension code into new file
